Refactor out codec for table format
init
commit_hash:65402bd8880a077306c1ded09b6d1aa8ea03cd1e
