+++ yql/essentials/minikql/protobuf_udf/ut/ya.make
@@ -11,6 +11,7 @@ SRCS(
 PEERDIR(
     yt/yql/providers/yt/lib/schema
     yt/yql/providers/yt/common
+    yt/yql/providers/yt/codec
     yql/essentials/public/udf/service/exception_policy
     yql/essentials/minikql
     yql/essentials/public/udf
