+++ yql/essentials/providers/common/codec/ya.make
@@ -19,7 +19,6 @@ PEERDIR(
     library/cpp/yson
     library/cpp/json
     library/cpp/enumbitset
-    yt/yt/library/decimal
 )
 
 YQL_LAST_ABI_VERSION()
