+++ yql/essentials/providers/common/codec/yql_codec.h
@@ -63,42 +63,14 @@ struct TCodecContext {
 void SkipYson(char cmd, TInputBuf& buf);
 void CopyYson(char cmd, TInputBuf& buf, TVector<char>& yson);
 void CopyYsonWithAttrs(char cmd, TInputBuf& buf, TVector<char>& yson);
-NKikimr::NUdf::TUnboxedValue ReadYsonValue(NKikimr::NMiniKQL::TType* type,  ui64 nativeYtTypeFlags, const NKikimr::NMiniKQL::THolderFactory& holderFactory, char cmd, TInputBuf& buf, bool isTableFormat);
+TStringBuf ReadNextString(char cmd, TInputBuf& buf);
+NKikimr::NUdf::TUnboxedValue ReadYsonValue(NKikimr::NMiniKQL::TType* type, const NKikimr::NMiniKQL::THolderFactory& holderFactory, char cmd, TInputBuf& buf);
 
 TMaybe<NKikimr::NUdf::TUnboxedValue> ParseYsonValue(const NKikimr::NMiniKQL::THolderFactory& holderFactory,
-    const TStringBuf& yson, NKikimr::NMiniKQL::TType* type,  ui64 nativeYtTypeFlags, IOutputStream* err, bool isTableFormat);
-TMaybe<NKikimr::NUdf::TUnboxedValue> ParseYsonNode(const NKikimr::NMiniKQL::THolderFactory& holderFactory,
-    const NYT::TNode& node, NKikimr::NMiniKQL::TType* type,  ui64 nativeYtTypeFlags, IOutputStream* err);
+    const TStringBuf& yson, NKikimr::NMiniKQL::TType* type, IOutputStream* err);
 
 TMaybe<NKikimr::NUdf::TUnboxedValue> ParseYsonNodeInResultFormat(const NKikimr::NMiniKQL::THolderFactory& holderFactory,
     const NYT::TNode& node, NKikimr::NMiniKQL::TType* type, IOutputStream* err);
-
-extern "C" void ReadYsonContainerValue(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags,
-    const NKikimr::NMiniKQL::THolderFactory& holderFactory, NKikimr::NUdf::TUnboxedValue& value, NCommon::TInputBuf& buf,
-    bool wrapOptional);
-
-void SkipSkiffField(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags, TInputBuf& buf);
-
-NKikimr::NUdf::TUnboxedValue ReadSkiffNativeYtValue(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags,
-    const NKikimr::NMiniKQL::THolderFactory& holderFactory, TInputBuf& buf);
-
-NKikimr::NUdf::TUnboxedValue ReadSkiffData(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags, NCommon::TInputBuf& buf);
-extern "C" void ReadContainerNativeYtValue(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags,
-    const NKikimr::NMiniKQL::THolderFactory& holderFactory, NKikimr::NUdf::TUnboxedValue& value, NCommon::TInputBuf& buf,
-    bool wrapOptional);
-
-extern "C" void WriteYsonContainerValue(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags,
-    const NKikimr::NUdf::TUnboxedValuePod& value, NCommon::TOutputBuf& buf);
-
-void WriteSkiffData(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags, const NKikimr::NUdf::TUnboxedValuePod& value, NCommon::TOutputBuf& buf);
-
-void WriteSkiffNativeYtValue(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags,
-    const NKikimr::NUdf::TUnboxedValuePod& value, NCommon::TOutputBuf& buf);
-
-extern "C" void WriteContainerNativeYtValue(NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags,
-    const NKikimr::NUdf::TUnboxedValuePod& value, NCommon::TOutputBuf& buf);
-
-void WriteYsonValueInTableFormat(TOutputBuf& buf, NKikimr::NMiniKQL::TType* type, ui64 nativeYtTypeFlags, const NKikimr::NUdf::TUnboxedValuePod& value, bool topLevel);
 
 TExprNode::TPtr ValueToExprLiteral(const TTypeAnnotationNode* type, const NKikimr::NUdf::TUnboxedValuePod& value, TExprContext& ctx,
     TPositionHandle pos = {});
